refactoring, linting, the good stuff
