+++ src/components/Header/Header.js
@@ -1,22 +1,13 @@
 import React, { Component } from "react";
-// import { Link } from 'react-router-dom';
-import "../styles/Header.css";
+import "./Header.css";
 
 import {
   Alignment,
   Button,
   AnchorButton,
   Navbar,
-  Tag,
-  Popover,
-  Position,
-  Tabs,
-  Tab,
-  Icon
 } from "@blueprintjs/core";
 
-import UserMenu from "./UserMenu";
-import Modal from "../components/Modal/Modal";
 class Header extends Component {
   constructor() {
     super();
@@ -47,34 +38,19 @@ class Header extends Component {
     });
   }
 
-  toggleProfileModal() {
-    return this.state.profileIsOpen ? this.setState({ profileIsOpen: false }) : this.setState({ profileIsOpen: true })
-  }
-
   render() {
-    const {keys, loadingKeys, profileIsOpen} = this.state
-    const { stats, user, auth, callAction } = this.props;
-    const baseProps = {
-      content: <UserMenu auth={auth} toggleProfileModal={this.toggleProfileModal.bind(this)}/>,
-      position: Position.BOTTOM_RIGHT
-    };
+    const {keys, loadingKeys} = this.state
+    const { user, auth } = this.props;
     return (      
       <Navbar
       // fixedToTop="true"
       >
-        <Modal callAction={callAction} isOpen={profileIsOpen} toggleOverlay={this.toggleProfileModal.bind(this)} user={user} />
         <Navbar.Group align={Alignment.LEFT}>
           <Navbar.Heading className="Header-logo">
             {/* <b> VGO Cases </b> */}
-            <img src="https://i.imgur.com/csrxCfC.png" />
+            <img alt="VGODOGG LOGO" src="https://i.imgur.com/csrxCfC.png" />
           </Navbar.Heading>
           <Navbar.Divider />
-          {/* <AnchorButton 
-            href="/#/"
-            className="bp3-minimal" 
-            icon="home" 
-            text="Home" 
-          /> */}
           <AnchorButton
             href="/#/cases"
             className="bp3-minimal"
@@ -95,15 +71,6 @@ class Header extends Component {
           />
         </Navbar.Group>
         <Navbar.Group align={Alignment.RIGHT}>
-          {/* <Navbar.Divider /> */}
-          {/* <Tag
-            minimal={true}
-            interactive={true}
-            large={true}
-            icon="key"
-            onClick={this.updateKeyCount.bind(this)}
-            text={ this.state.keys }
-          /> */}
           <Button
             minimal={true}
             icon="key"
@@ -111,7 +78,6 @@ class Header extends Component {
             text={keys}
             loading={loadingKeys}
           />
-          {/* <Button className="bp3-minimal" icon="notifications" text="" /> */}
           {!user ? (
             <Button
               // className="bp3-minimal"
@@ -124,16 +90,7 @@ class Header extends Component {
                 className="bp3-minimal"
                 icon="person"
                 text={user.username}
-                // rightIcon="caret-down"
               />
-            // <Popover {...baseProps} minimal={true}>
-            //   <Button
-            //     className="bp3-minimal"
-            //     icon="person"
-            //     text={user.username}
-            //     rightIcon="caret-down"
-            //   />
-            // </Popover>
           )}
         </Navbar.Group>
       </Navbar>
