+++ src/components/OpeningFeed/OpeningFeed.css
