+++ src/pages/SpinCase/SpinCase.css
